Broke off a few more tests
