+++ test/HiGlassComponentTest.js
@@ -22,19 +22,15 @@ import {
   getTiledPlot,
   waitForJsonComplete,
   waitForTilesLoaded,
-  waitForTransitionsFinished,
 } from '../app/scripts/utils';
 
 // View configs
 import {
   // paperFigure1,
-  osmConf,
-  geneAnnotationsOnly,
   annotationsTilesView,
   horizontalAndVerticalMultivec,
   divergentTrackConfig,
   simpleCenterViewConfig,
-  rectangleDomains,
   project1D,
   chromInfoTrack,
   heatmapTrack,
@@ -43,9 +39,7 @@ import {
   oneZoomedOutViewConf,
   // valueIntervalTrackViewConf,
   chromosomeGridTrack,
-  testViewConfX1,
   testViewConfX2,
-  restrictedZoom,
 } from './view-configs';
 
 configure({ adapter: new Adapter() });
@@ -64,394 +58,6 @@ describe('Simple HiGlassComponent', () => {
 
   afterAll(async () => {
     await fetchMockHelper.storeDataAndResetFetchMock();
-  });
-
-  describe('API tests', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'width:800px;background-color: lightgreen');
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(
-        <HiGlassComponent options={{ bounded: false }} viewConfig={osmConf} />,
-        { attachTo: div },
-      );
-
-      hgc.update();
-      waitForTilesLoaded(hgc.instance(), done);
-      // done();
-    });
-
-    it('Switches to the osm tiles track', () => {
-      const { views } = hgc.instance().state;
-      // console.log('views:', views);
-
-      const view = views.aa;
-
-      view.tracks.center[0].type = 'osm-2d-tile-ids';
-      view.tracks.center[0].uid = 'bb';
-
-      hgc.setState({
-        views,
-      });
-    });
-  });
-
-  describe('API tests', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'width:800px;background-color: lightgreen');
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: false }}
-          viewConfig={geneAnnotationsOnly}
-        />,
-        { attachTo: div },
-      );
-
-      hgc.update();
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Zooms to a location', (done) => {
-      hgc.instance().zoomTo('aa', 1, 1000000);
-
-      waitForTransitionsFinished(hgc.instance(), () => {
-        const svgText = hgc.instance().createSVGString();
-
-        // make sure none of the chromosome labels are left
-        // over after zooming
-        expect(svgText.indexOf('chr11')).toEqual(-1);
-
-        // hgc.instance().handleExportSVG();
-        done();
-      });
-    });
-
-    it('Zooms a little closer', (done) => {
-      hgc.instance().zoomTo('aa', 165061, 945306);
-
-      waitForTransitionsFinished(hgc.instance(), () => {
-        done();
-      });
-    });
-  });
-
-  describe('Zoom restriction tests', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'width:800px;background-color: lightgreen');
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: false }}
-          viewConfig={restrictedZoom}
-        />,
-        { attachTo: div },
-      );
-
-      hgc.update();
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Has the corrent limits', (done) => {
-      const zoomLimits = hgc.instance().tiledPlots.aa.props.zoomLimits;
-      expect(zoomLimits[0]).toEqual(0.002);
-      expect(zoomLimits[1]).toEqual(2);
-      done();
-    });
-
-    it('Zooms in and respects zoom limit', (done) => {
-      // Create a wheel event that zooms in beying the zoom limit
-      const evt = new WheelEvent('wheel', {
-        deltaX: 0,
-        deltaY: -500,
-        deltaZ: 0,
-        deltaMode: 0,
-        clientX: 262,
-        clientY: 572,
-        screenX: 284,
-        screenY: 696,
-        view: window,
-        bubbles: true,
-        cancelable: true,
-      });
-
-      hgc.instance().tiledPlots.aa.trackRenderer.element.dispatchEvent(evt);
-
-      waitForTransitionsFinished(hgc.instance(), () => {
-        // Make sure, it does not zoom too far
-        const k = hgc.instance().tiledPlots.aa.trackRenderer.zoomTransform.k;
-        expect(k).toEqual(2);
-
-        done();
-      });
-    });
-  });
-
-  describe('2D Rectangle Annotations', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'width:800px;background-color: lightgreen');
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: false }}
-          viewConfig={rectangleDomains}
-        />,
-        { attachTo: div },
-      );
-
-      hgc.update();
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Check to make sure that the rectangles are initially small', (done) => {
-      let track = getTrackObjectFromHGC(hgc.instance(), 'aa', 'rectangles1');
-
-      let hasSmaller = false;
-      for (const uid of Object.keys(track.drawnRects)) {
-        if (track.drawnRects[uid].width < 5) {
-          hasSmaller = true;
-          break;
-        }
-      }
-
-      expect(hasSmaller).toEqual(true);
-
-      const { views } = hgc.instance().state;
-      track = getTrackByUid(views.aa.tracks, 'rectangles1');
-
-      track.options.minSquareSize = '8';
-
-      hgc.setState({
-        views,
-      });
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Make sure that the rectangles are large', (done) => {
-      let track = getTrackObjectFromHGC(hgc.instance(), 'aa', 'rectangles1');
-
-      let hasSmaller = false;
-      for (const uid of Object.keys(track.drawnRects)) {
-        if (track.drawnRects[uid].width < 5) {
-          hasSmaller = true;
-          break;
-        }
-      }
-
-      expect(hasSmaller).toEqual(false);
-
-      const { views } = hgc.instance().state;
-      track = getTrackByUid(views.aa.tracks, 'rectangles1');
-
-      track.options.minSquareSize = '5';
-
-      hgc.setState({
-        views,
-      });
-
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Exports to SVG', () => {
-      hgc.instance().createSVG();
-    });
-  });
-
-  describe('Export SVG properly', () => {
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'width:800px;background-color: lightgreen');
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: false }}
-          viewConfig={testViewConfX1}
-        />,
-        { attachTo: div },
-      );
-
-      hgc.update();
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Exports to SVG', () => {
-      const svg = hgc.instance().createSVG();
-      const svgText = new XMLSerializer().serializeToString(svg);
-
-      expect(svgText.indexOf('rect')).toBeGreaterThan(0);
-      // hgc.instance().handleExportSVG();
-    });
-
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'width:800px;background-color: lightgreen');
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: false }}
-          viewConfig={project1D}
-        />,
-        { attachTo: div },
-      );
-
-      hgc.update();
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    // it('Exports to SVG', (done) => {
-    //   // const svg = hgc.instance().createSVG();
-    //   // const svgText = new XMLSerializer().serializeToString(svg);
-    //
-    //   // expect(svgText.indexOf('dy="-17"')).toBeGreaterThan(0);
-    //   // hgc.instance().handleExportSVG();
-    //
-    //   done();
-    // });
-
-    it('Replaces one of the views and tries to export again', () => {
-      let { views } = hgc.instance().state;
-
-      const newView = JSON.parse(JSON.stringify(views.aa));
-
-      hgc.instance().handleCloseView('aa');
-      ({ views } = hgc.instance().state);
-
-      newView.uid = 'a2';
-      newView.layout.i = 'a2';
-
-      views.a2 = newView;
-
-      hgc.instance().setState({ views });
-
-      // this used to raise an error because the hgc.instance().tiledPlots
-      // would maintain a reference to the closed view and we would try
-      // to export it as SVG
-      hgc.instance().createSVG();
-
-      // hgc.instance().createSVG();
-    });
-
-    it('Cleans up previously created instances and mounts a new component', (done) => {
-      if (hgc) {
-        hgc.unmount();
-        hgc.detach();
-      }
-
-      if (div) {
-        global.document.body.removeChild(div);
-      }
-
-      div = global.document.createElement('div');
-      global.document.body.appendChild(div);
-
-      div.setAttribute('style', 'width:800px;background-color: lightgreen');
-      div.setAttribute('id', 'simple-hg-component');
-
-      hgc = mount(
-        <HiGlassComponent
-          options={{ bounded: false }}
-          viewConfig={project1D}
-        />,
-        { attachTo: div },
-      );
-
-      hgc.update();
-      waitForTilesLoaded(hgc.instance(), done);
-    });
-
-    it('Exports to SVG', () => {
-      const svg = hgc.instance().createSVG();
-      const svgText = new XMLSerializer().serializeToString(svg);
-
-      // check to make sure that the horizontal labels shifted down
-      // the horizontal lines' labels should be shifted down
-      expect(svgText.indexOf('dy="14"')).toBeGreaterThan(0);
-
-      // check to make sure that chromosome tick labels are there
-      expect(svgText.indexOf('chr17: 40,500,000')).toBeGreaterThan(0);
-
-      // check to make sure that the chromosome ticks are present
-      expect(svgText.indexOf('line x1')).toBeGreaterThan(0);
-      expect(svgText.indexOf('#808080')).toBeGreaterThan(0);
-
-      // hgc.instance().handleExportSVG();
-    });
   });
 
   describe('Track type menu tests', () => {
